move things (part 1)
