+++ cvlr-solana/src/lib.rs
@@ -8,11 +8,6 @@ pub mod pubkey;
 pub mod token;
 
 pub use clock::*;
-#[cfg(feature = "rt")]
-pub use layout::global_database::{init, GlobalDatabase};
-pub use layout::{
-    cvlr_deserialize_nondet_accounts, cvlr_deserialize_nondet_accounts as cvlr_nondet_acc_infos,
-    cvlr_new_account_info, fun_acc_infos_with_mem_layout,
-};
 pub use log::*;
 pub use nondet::*;
+pub use layout::*;
